# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a service that allows you to receive data stored in the blockchain.",
        "title": "Blockchain Explorer",
        "contact": {},
        "version": "0.9.1"
    },
    "host": "localhost",
    "basePath": "/api",
    "paths": {
        "/api/auth/refresh": {
            "post": {
                "description": "refresh token pair",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh",
                "operationId": "refresh",
                "parameters": [
                    {
                        "description": "access token",
                        "name": "access_token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "refresh token",
                        "name": "refresh_token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TokenPair"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apperrors.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.Error"
                        }
                    }
                }
            }
        },
        "/api/auth/sign-ip": {
            "post": {
                "description": "Authorize user in the service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "SignIn",
                "operationId": "sign-in",
                "parameters": [
                    {
                        "description": "user login",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "user password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TokenPair"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apperrors.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperrors.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apperrors.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.Error"
                        }
                    }
                }
            }
        },
        "/api/auth/sign-out": {
            "post": {
                "description": "Logout from the service",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "SignOut",
                "operationId": "sign-out",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apperrors.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.Error"
                        }
                    }
                }
            }
        },
        "/api/auth/sign-up": {
            "post": {
                "description": "Register user in the service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "SignUp",
                "operationId": "sign-up",
                "parameters": [
                    {
                        "description": "user email/login",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "user password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "confirm user password",
                        "name": "confirm_password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperrors.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apperrors.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.Error"
                        }
                    }
                }
            }
        },
        "/v1/block/{block}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get block by hash or id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blocks"
                ],
                "summary": "GetBlock",
                "operationId": "get-block",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Block id or hash",
                        "name": "block",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Block"
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/blocks": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get blocks with limit and offset",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blocks"
                ],
                "summary": "GetBlocks",
                "operationId": "get-blocks",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the amount of blocks you want to get",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset from the beginning of the data in database",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Block"
                            }
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/transactions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get transactions with limit and offset params",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "GetTransactions",
                "operationId": "get-transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the amount of transactions you want to get",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset from the beginning of the data in database",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Transaction"
                            }
                        }
                    },
                    "404": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apperrors.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Arg": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Value"
                    }
                },
                "bytes": {
                    "type": "string"
                },
                "int": {
                    "type": "string"
                },
                "prim": {
                    "type": "string"
                }
            }
        },
        "models.BigMapDiff": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "big_map": {
                    "type": "string"
                },
                "key": {
                    "$ref": "#/definitions/models.Value"
                },
                "key_hash": {
                    "type": "string"
                }
            }
        },
        "models.Block": {
            "type": "object",
            "properties": {
                "chain_id": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "header": {
                    "$ref": "#/definitions/models.Header"
                },
                "metadata": {
                    "$ref": "#/definitions/models.BlockMetadata"
                },
                "operations": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/models.Operation"
                        }
                    }
                },
                "protocol": {
                    "type": "string"
                }
            }
        },
        "models.BlockMetadata": {
            "type": "object",
            "properties": {
                "baker": {
                    "type": "string"
                },
                "consumed_gas": {
                    "type": "string"
                },
                "level": {
                    "$ref": "#/definitions/models.LevelInfo"
                },
                "level_info": {
                    "$ref": "#/definitions/models.LevelInfo"
                }
            }
        },
        "models.Content": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "counter": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "fee": {
                    "type": "string"
                },
                "gas_limit": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/models.Metadata"
                },
                "parameters": {
                    "$ref": "#/definitions/models.Parameters"
                },
                "source": {
                    "type": "string"
                },
                "storage_limit": {
                    "type": "string"
                }
            }
        },
        "models.Diff": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "updates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Update"
                    }
                }
            }
        },
        "models.Header": {
            "type": "object",
            "properties": {
                "baker_fee": {
                    "type": "integer"
                },
                "fitness": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "operations_hash": {
                    "type": "string"
                },
                "predecessor": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "proto": {
                    "type": "integer"
                },
                "signature": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "validation_pass": {
                    "type": "integer"
                }
            }
        },
        "models.InternalOperationResult": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "nonce": {
                    "type": "integer"
                },
                "result": {
                    "$ref": "#/definitions/models.Result"
                },
                "source": {
                    "type": "string"
                }
            }
        },
        "models.LazyStorageDiff": {
            "type": "object",
            "properties": {
                "diff": {
                    "$ref": "#/definitions/models.Diff"
                },
                "id": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                }
            }
        },
        "models.LevelInfo": {
            "type": "object",
            "properties": {
                "cycle": {
                    "type": "integer"
                },
                "cycle_position": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                }
            }
        },
        "models.Metadata": {
            "type": "object",
            "properties": {
                "balance_updates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MetadataBalanceUpdate"
                    }
                },
                "internal_operation_results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.InternalOperationResult"
                    }
                },
                "operation_result": {
                    "$ref": "#/definitions/models.OperationResult"
                }
            }
        },
        "models.MetadataBalanceUpdate": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "change": {
                    "type": "string"
                },
                "contract": {
                    "type": "string"
                },
                "cycle": {
                    "type": "integer"
                },
                "delegate": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                }
            }
        },
        "models.Operation": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "chain_id": {
                    "type": "string"
                },
                "contents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Content"
                    }
                },
                "hash": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                }
            }
        },
        "models.OperationResult": {
            "type": "object",
            "properties": {
                "big_map_diff": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BigMapDiff"
                    }
                },
                "consumed_gas": {
                    "type": "string"
                },
                "consumed_milligas": {
                    "type": "string"
                },
                "lazy_storage_diff": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.LazyStorageDiff"
                    }
                },
                "status": {
                    "type": "string"
                },
                "storage": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Storage"
                    }
                },
                "storage_size": {
                    "type": "string"
                }
            }
        },
        "models.Parameters": {
            "type": "object",
            "properties": {
                "entrypoint": {
                    "type": "string"
                },
                "value": {
                    "$ref": "#/definitions/models.Value"
                }
            }
        },
        "models.Result": {
            "type": "object",
            "properties": {
                "balance_updates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ResultBalanceUpdate"
                    }
                },
                "consumed_gas": {
                    "type": "string"
                },
                "consumed_milligas": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.ResultBalanceUpdate": {
            "type": "object",
            "properties": {
                "change": {
                    "type": "string"
                },
                "contract": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                }
            }
        },
        "models.Storage": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Arg"
                    }
                },
                "int": {
                    "type": "string"
                },
                "prim": {
                    "type": "string"
                }
            }
        },
        "models.TokenPair": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "models.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "block_hash": {
                    "type": "string"
                },
                "branch": {
                    "type": "string"
                },
                "consumed_milligas": {
                    "type": "string"
                },
                "counter": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "fee": {
                    "type": "string"
                },
                "gas_limit": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "storage_size": {
                    "type": "string"
                }
            }
        },
        "models.Update": {
            "type": "object",
            "properties": {
                "key": {
                    "$ref": "#/definitions/models.Value"
                },
                "key_hash": {
                    "type": "string"
                }
            }
        },
        "models.Value": {
            "type": "object",
            "properties": {
                "int": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization (Bearer token)",
            "in": "header"
        }
    }
}